requirements.txt
pymupdf==1.24.1
pdfplumber==0.11.0

langchain==0.2.16
langchain-chroma==0.1.4

chromadb==0.5.23

sentence-transformers==2.7.0
transformers==4.40.2
huggingface-hub==0.24.7
tokenizers==0.19.1

torch==2.3.1
numpy<2.0.0

anthropic

fastapi==0.115.0
uvicorn==0.30.6

streamlit==1.38.0

ragas==0.1.21
datasets==2.21.0

python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.8.2
pydantic-settings==2.4.0
tqdm==4.66.5